Let's marshall encoder button press events to main loop()
This is so that we can do things that consume CPU time if we want without the watchdog being upset. Also, more kosher to serialize all the main logic.

Note: The ESP32 way to do this is using FreeRTOS queues, but so far this project has been not ESP32-specific.